A new version of the add-on with two additional features:
Added ability to export to csv
As part of this change, the add-on now requires the Apache Commons IO
library (http://commons.apache.org/io/). This was a little trickier to
incorporate than expected, but is based on the following code mentioned
in the POI quick guide:

/src/examples/src/org/apache/poi/hssf/eventusermodel/examples/XLS2CSVmra.java

Hooking it up was the tricky part, because I'm not particularly strong
on Java input / output streams.

There's a new class CsvExport which extends ExcelExport which should be
used to get CSV output. The demo application has an example of this.

It seems to work well, but any limitations are limitations in
XLS2CSVmra.java, which is open source and incorporated into the
project--i.e. for now, bugfixes aside, if anyone wants improvements to
this, they should join the project and implement them themselves.



Added ability to format cells based on Table's formatPropertyValue()
Since Table protects its formatted properties, the user is required to
have their Table implement ExportableFormattedProperty which just makes
formatted properties accessible.

There is also a utility class called PropertyFormatTable that has an
implementation of ExportableFormattedProperty. A user can just extend
this class instead of Table. The demo application also has an example of
this.