Fixed two reported bugs -
1) Inefficient use of cell styles which generated exception when a POI
limit of 4000 per sheet was reached.
CellStyles now handled a lot more efficiently. There is now a separate
dateDataCellStyle and a doubleDataCellStyle.

2) Nullpointer exception with null property and using Table
formatPropertyValue
This was a result of the following issue.  We identify whether a cell
should use formatPropertyValue or not by whether formatPropertyValue
returns something different than just property.toString().  If yes, then
we use the formatPropertyValue format, but it is always a String.  If
no, then we use the regular property value, and it is set properly
according to the data type.  A problem happens when we have a null when
we can't do this check.  So now, we just use formatPropertyValue (if the
use formatPropertyValue flag is set) if it is null and we end up with an
empty string instead of a null date or something like that.  So, if
there is a null property in a column of dates, there will be an empty
String cell in the Excel column.  This is probably fine.

One side-effect I noticed of enabling formatPropertyValue is that, when
used to format numbers, it causes the totals cell not to work, because
the totals formula is summing up Strings.