Enabled exporting multiple tables in one file.  Create an ExcelExport with a Table and call convertTable() instead of export().  Then call setNextTable with the next table and the next sheet name.  Then call convertTable() again.  Do as many times as you want, then on the last sheet call export() (which just calls convertTable() followed by sendConverted().
Also enabled using a SXSSFWorkbook. There is now a constructor where you
can pass in a workbook.  If you don't pass in one, it just creates a
new HSSFWorkbook, as before.  However, you can pass in a SXSSFWorkbook
and it will use that.  Was unable to test because SXSSFWorkbook does not
yet come with the POI jar distribution.