1) Main change is that I Incorporated Thomas Trinn's enhancements making things work with FilteringTable (see comments below). 2) Mavenized the project 3) Fixed the typo pointed out by Okihara Herbst on the form ("cvs" instead of "csv") 4) Removed some deprecated references.  One remains that still needs to be refactored out. 5) Slight changes in the versions used for poi (from 3.9 with yyyymmdd to final 3.9) and commons-io (from 2.1 to 2.2)
-----

Thomas's comments RE how the enhancements work: It does not work with the FilteringTable addon. I downloaded the source code, and I saw I could make it work with CustomTable. And so I did.
I hope you could take a look at it: table export path for CustomTable
And it would be really great if we could push it up in a new version.

The things I made:
creating an interface: TableHolder (it has all the specific methods witch are Table dependent)
creating an implementation of TableHolder: DefaultTableHolder (for com.vaadin.ui.Table)
changed the TableExport, ExcelExport and CsvExport classes to work with TableHolder
Basically if somebody needs only the com.vaadin.ui.Table, he/she does not have to do any special stuff. In case somebody needs it for com.vaadin.ui.CustomTable, he/she have to create an implementation of the TableHolder and create the ExcelExport object by passing that at the constructor. So your project does not depend on the filtering table addon, but it is easy to use with it.

In the attachment you see a CustomTableHolder the implementation of TableHolder for CustomTable and CustomTableExportableColumnGenerator the ColumnGenerator for CustomTable.