DefaultTableHolder now accepts Grid and Container (so it is a TableOrGridOrContainerHolder) although the name hasn't changed.  The constructor to classes like ExcelExport accept a Table or TableHolder. If you passed in a Table, it created a TableHolder for that Table behind the scenes and queried the TableHolder for the data.  So now if you pass a TableHolder holding a Grid or Container and pass it into ExcelExport, it will work as well.
The demo UI has been updated to have an example of a Grid with a GeneratedPropertyContainer.

Some additional behind the scenes changes:

In adding support for Grid, had to completely redo the Maven project setup.  Was running into weird NoClassDefFoundError and ClassNotFoundException errors relating to Grid.  Turned out the something was messed up with the Intellij project setup. Despite having a late version of Vaadin 7.5.10 in the pom, it was using a pre-Grid version of Vaadin when running the test application. Re-organized the project on the latest Vaadin Maven archetype. Was a 2 day ordeal. Boy that archetype is finicky. Turns out it's really sensitive to whether you use a @WebServlet annotation instead of a web.xml file. The @WebServlet annotation works a lot better with the new setup.

Also the test app now needs widgetset compilation which makes building slower.  Not sure exactly why this is required.  Possibly because FilteringTable is a dependency although it was a dependency before. Trying to get add-on packaging working again with this new archetype and excluding the 23 meg of widgetset files (which aren't needed for the add-on, only for running the test UI) was another 2 days of fun with Maven minutia.